+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#### 1.0.3
+* Fixed an error while accepting ready check in combat
+* Update deprecated API for accepting ready check
 #### 1.0.2
 Removed unused files
 #### 1.0.1